# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: SOL Duel Game API
  description: |
    **SOL Duel** is a real-time PvP game built on the Solana blockchain. This API provides endpoints for player authentication, match management, game state synchronization, and blockchain interactions.

    ## Features
    - 🔐 Wallet-based authentication with signature verification
    - ⚔️ Real-time multiplayer combat system
    - 💰 Solana-based wagering and rewards
    - 🏆 Tournament and leaderboard systems
    - 🔗 WebSocket real-time updates
    - 📊 Comprehensive player statistics

    ## Authentication
    This API uses JWT tokens for authentication. Players authenticate using their Solana wallet signature.

    ## Rate Limiting
    - 100 requests per 15 minutes per IP address
    - WebSocket connections limited to 10 per authenticated user

    ## WebSocket Events
    Real-time game events are handled through WebSocket connections. See the WebSocket documentation for details.
  version: 1.0.0
  contact:
    name: SOL Duel API Support
    email: support@solduel.game
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.solduel.game
    description: Production server
  - url: https://staging-api.solduel.game
    description: Staging server
  - url: http://localhost:5000
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /api/games/{id}/join:
    post:
      tags:
        - Games
      summary: Join game
      description: Join an existing game as a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Game ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - wagerAmount
              properties:
                wagerAmount:
                  type: number
                  minimum: 0
                  description: Wager amount in SOL (must match game requirements)
                  example: 0.1
      responses:
        '200':
          description: Successfully joined game
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Successfully joined game"
                  data:
                    type: object
                    properties:
                      game:
                        $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or wallet verification
  schemas:
...
    Game:
      type: object
      properties:
        id:
          type: string
          description: Unique game identifier
          example: "game_12345"
        gameType:
          type: string
          enum: [duel, tournament, practice]
          example: "duel"
        status:
          type: string
          enum: [waiting, active, finished, cancelled]
          example: "waiting"
        wagerAmount:
          type: number
          description: Wager amount in SOL
          example: 0.1
        maxPlayers:
          type: integer
          example: 2
        currentPlayers:
          type: integer
          description: Number of players currently in game
          example: 1
        timeLimit:
          type: integer
          description: Time limit in seconds
          example: 300
        isPrivate:
          type: boolean
          example: false
        creator:
          type: object
          properties:
            id:
              type: string
            username:
              type: string
            eloRating:
              type: integer
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: string
          description: Additional error details (optional)
        timestamp:
          type: string
          format: date-time
  responses:
    ValidationError:
      description: Validation error in request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Validation Error"
            message: "Invalid request data"
            details: "username must be between 3 and 30 characters"
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Authentication Failed"
            message: "Valid authentication token required"
    ForbiddenError:
      description: Access denied - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Access Denied"
            message: "You do not have permission to perform this action"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Not Found"
            message: "The requested resource was not found"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal Server Error"
            message: "An unexpected error occurred"
...